requirements.txt
scrapy==2.8.0
python-dotenv==1.0.0
psycopg[binary]==3.1.8
